pyproject.toml
[tool.poetry]
name = "stock-crawler"
version = "0.1.0"
description = ""
authors = ["LittleWhiteYA <f46284628@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
beautifulsoup4 = "^4.9.3"
pymongo = "^3.11.3"
python-dotenv = "^0.17.0"
python-dateutil = "^2.8.1"
flake8 = "^3.9.0"
pandas = "^1.2.4"
jupyter = "^1.0.0"
matplotlib = "^3.4.1"
tabulate = "^0.8.9"
black = "^21.12b0"
pydash = "^5.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
reportOptionalSubscript = false
typeCheckingMode = "off"
